Add initial d3 support
