+++ di_website/visualisation/static/visualisation/js/dicharts.js
@@ -1,1 +1,1 @@
-var DICharts=function(t){var e={};function n(r){if(e[r])return e[r].exports;var o=e[r]={i:r,l:!1,exports:{}};return t[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}return n.m=t,n.c=e,n.d=function(t,e,r){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:r})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var o in t)n.d(r,o,function(e){return t[e]}.bind(null,o));return r},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=1)}([function(t,e,n){"use strict";var r=function(t){return function(t){return!!t&&"object"==typeof t}(t)&&!function(t){var e=Object.prototype.toString.call(t);return"[object RegExp]"===e||"[object Date]"===e||function(t){return t.$$typeof===o}(t)}(t)};var o="function"==typeof Symbol&&Symbol.for?Symbol.for("react.element"):60103;function i(t,e){return!1!==e.clone&&e.isMergeableObject(t)?f((n=t,Array.isArray(n)?[]:{}),t,e):t;var n}function a(t,e,n){return t.concat(e).map((function(t){return i(t,n)}))}function c(t){return Object.keys(t).concat(function(t){return Object.getOwnPropertySymbols?Object.getOwnPropertySymbols(t).filter((function(e){return t.propertyIsEnumerable(e)})):[]}(t))}function u(t,e){try{return e in t}catch(t){return!1}}function l(t,e,n){var r={};return n.isMergeableObject(t)&&c(t).forEach((function(e){r[e]=i(t[e],n)})),c(e).forEach((function(o){(function(t,e){return u(t,e)&&!(Object.hasOwnProperty.call(t,e)&&Object.propertyIsEnumerable.call(t,e))})(t,o)||(u(t,o)&&n.isMergeableObject(e[o])?r[o]=function(t,e){if(!e.customMerge)return f;var n=e.customMerge(t);return"function"==typeof n?n:f}(o,n)(t[o],e[o],n):r[o]=i(e[o],n))})),r}function f(t,e,n){(n=n||{}).arrayMerge=n.arrayMerge||a,n.isMergeableObject=n.isMergeableObject||r,n.cloneUnlessOtherwiseSpecified=i;var o=Array.isArray(e);return o===Array.isArray(t)?o?n.arrayMerge(t,e,n):l(t,e,n):i(e,n)}f.all=function(t,e){if(!Array.isArray(t))throw new Error("first argument should be an array");return t.reduce((function(t,n){return f(t,n,e)}),{})};var s=f;t.exports=s},function(t,e,n){"use strict";n.r(e),n.d(e,"handler",(function(){return S})),n.d(e,"Chart",(function(){return i}));function r(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}function o(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var i=function(){function t(e){var n=this;!function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,t),o(this,"chartElement",void 0),o(this,"theme","default"),o(this,"showLoading",(function(){var t=n.getParentElement();t&&t.classList.add("chart-container--loading")})),o(this,"hideLoading",(function(){var t=n.getParentElement();t&&function(t){t.classList.remove("chart-container--loading");var e=t.querySelector(".chart-loading");e&&e.classList.add("hide")}(t)})),o(this,"addFilter",(function(t,e,r){var o=n.getElement(t);return e.forEach((function(t){var e=document.createElement("option");e.value=t,e.text=n.getFilterText(t,r),o.appendChild(e)})),o.classList.add("data-selector--active"),o})),o(this,"getThemes",(function(){return{rainbow:["#e84439","#eb642b","#f49b21","#109e68","#0089cc","#893f90","#c2135b","#f8c1b2","#f6bb9d","#fccc8e","#92cba9","#88bae5","#c189bb","#e4819b"],default:["#6c120a","#a21e25","#cd2b2a","#dc372d","#ec6250","#f6b0a0","#fbd7cb","#fce3dc"],sunflower:["#7d4712","#ba6b15","#df8000","#f7a838","#fac47e","#fedcab","#fee7c1","#feedd4"],marigold:["#7a2e05","#ac4622","#cb5730","#ee7644","#f4a57c","#facbad","#fcdbbf","#fde5d4"],rose:["#65093d","#8d0e56","#9f1459","#d12568","#e05c86","#f3a5b6","#f6b8c1","#f9cdd0"],lavendar:["#42184c","#632572","#732c85","#994d98","#af73ae","#cb98c4","#deb5d6","#ebcfe5"],bluebell:["#0a3a64","#00538e","#1060a3","#4397d3","#77adde","#a3c7eb","#bcd4f0","#d3e0f4"],leaf:["#08492f","#005b3e","#00694a","#3b8c62","#74bf93","#a2d1b0","#b1d8bb","#c5e1cb"]}})),o(this,"getTheme",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:"default";return t?[t,n.getThemes()[t]]:[n.theme,n.getThemes()[n.theme]]})),this.chartElement=this.getElement(e)}var e,n,i;return e=t,(n=[{key:"getParentElement",value:function(){return this.chartElement.parentElement}},{key:"getFilterText",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},n=e.labelPrefix,r=e.labelSuffix;return"".concat(n||"").concat(t).concat(r||"")}},{key:"getElement",value:function(t){if(!t)throw new Error("An id to the element or an actual element is required");if("string"==typeof t){var e=document.getElementById(t);if(!e)throw new Error("No element with id ".concat(t));return e}return t}}])&&r(e.prototype,n),i&&r(e,i),t}(),a=n(0),c=n.n(a);function u(t){return(u="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function l(t,e){return function(t){if(Array.isArray(t))return t}(t)||function(t,e){if("undefined"==typeof Symbol||!(Symbol.iterator in Object(t)))return;var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=t[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!e||n.length!==e);r=!0);}catch(t){o=!0,i=t}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(t,e)||function(t,e){if(!t)return;if("string"==typeof t)return f(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return f(t,e)}(t,e)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function f(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}function s(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function d(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?s(Object(n),!0).forEach((function(e){m(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):s(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function y(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}function h(t,e){return(h=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function p(t){var e=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],(function(){}))),!0}catch(t){return!1}}();return function(){var n,r=v(t);if(e){var o=v(this).constructor;n=Reflect.construct(r,arguments,o)}else n=r.apply(this,arguments);return b(this,n)}}function b(t,e){return!e||"object"!==u(e)&&"function"!=typeof e?g(t):e}function g(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}function v(t){return(v=Object.setPrototypeOf?Object.getPrototypeOf:function(t){return t.__proto__||Object.getPrototypeOf(t)})(t)}function m(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var w,O,j={hovermode:"closest"},P={displayModeBar:!0,responsive:!0,showLink:!1,plotlyServerURL:"https://chart-studio.plotly.com",toImageButtonOptions:{width:1200,height:657},modeBarButtonsToRemove:["pan2d","select2d","lasso2d","hoverClosestCartesian","hoverCompareCartesian","toggleSpikelines"]},E=function(t){!function(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),e&&h(t,e)}(i,t);var e,n,r,o=p(i);function i(t,e){var n;return function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,i),m(g(n=o.call(this,t)),"options",void 0),m(g(n),"data",[]),m(g(n),"layout",{}),m(g(n),"config",{}),m(g(n),"plot",void 0),m(g(n),"sourceData",void 0),m(g(n),"filters",[]),m(g(n),"getPlot",(function(){return n.plot})),m(g(n),"setData",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[];return n.data=t,g(n)})),m(g(n),"setLayout",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return n.layout=c()(n.layout,t),n.setTheme(n.options.theme)})),m(g(n),"getLayout",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return c()(j,t)})),m(g(n),"setTheme",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:"default";return n.layout=d(d({},n.layout),{},{colorway:n.getThemes()[t]}),n.theme=t,g(n)})),m(g(n),"setConfig",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return n.config=c()(n.config,t),g(n)})),m(g(n),"getConfig",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return c()(P,t)})),m(g(n),"getOptions",(function(){return n.options})),m(g(n),"setFilters",(function(t){n.filters=t})),m(g(n),"getFilters",(function(){return n.filters})),m(g(n),"csv",(function(t){return new Promise((function(e){window.Plotly.d3.csv(t,(function(t){e(t)}))}))})),m(g(n),"updatePlot",(function(){return n.plot?window.Plotly.react(n.chartElement,n.data,n.getLayout(n.layout),n.getConfig(n.config)).then((function(t){return n.addPlot(t),n.options.theme||n.updateLayoutColorway(),n.hideLoading(),{plot:t,config:n.options}})):n.newPlot(n.chartElement,n.data,n.getLayout(n.layout),n.getConfig(n.config))})),m(g(n),"updateLayoutColorway",(function(){if(n.plot){var t=n.getThemes(),e=void 0;if(n.plot.calcdata.length>8)e=t.rainbow;else for(var r=document.body.classList,o=0,i=Object.entries(t);o<i.length;o++){var a=l(i[o],2),c=a[0],u=a[1];if(r.contains("body--".concat(c))&&u.length<=8){e=u;break}}e&&window.Plotly.relayout(n.chartElement,{colorway:e})}})),n.options=e,n}return e=i,(n=[{key:"setSourceData",value:function(t){return this.sourceData=t,this}},{key:"getSourceData",value:function(){return this.sourceData}},{key:"addPlot",value:function(t){return this.plot=t,this}},{key:"newPlot",value:function(t,e){var n=this,r=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},o=arguments.length>3&&void 0!==arguments[3]?arguments[3]:{};return window.Plotly.newPlot(t,e,r,this.getConfig(o)).then((function(t){return t.on("plotly_click",(function(t){n.options.onClick&&n.options.onClick(t,n)})),n.plot||n.initCustomWidgets(),n.addPlot(t),n.initPlotlyWidgets(),n.options.theme||n.updateLayoutColorway(),n.hideLoading(),{plot:t,config:n.options}}))}},{key:"initCustomWidgets",value:function(){var t=this,e=this.options.widgets,n=void 0===e?{}:e;n.filters&&n.filters.forEach((function(e){return t.handleFilter(e)}))}},{key:"initPlotlyWidgets",value:function(){var t=this.options.widgets,e=void 0===t?{}:t;e.legend&&this.handleLegend(e.legend)}},{key:"handleFilter",value:function(t){var e=this.getParentElement();if(e){var n=e.querySelector(".".concat(t.className));if(!n)throw new Error("No element with class ".concat(t.className));var r=t.options;!r&&t.getOptions&&(r=t.getOptions(this)),r&&r.length&&(this.addFilter(n,r||[]),this.addFilterEvents(n,t),this.setFilters(this.filters.concat({name:t.className,value:[r[0]]})))}}},{key:"addFilterEvents",value:function(t,e){var n=this;e.onChange&&t.addEventListener("change",(function(t){var r=t.target.value;n.setFilters(n.filters.map((function(t){return t.name===e.className&&-1===t.value.indexOf(r)&&(t.value=e.multi?t.value.concat(r):[r]),t}))),e.onChange(t,n)}))}},{key:"handleLegend",value:function(t){var e=this;this.plot&&t.onClick&&this.plot.on("plotly_legendclick",(function(n){t.onClick&&t.onClick(n,e)}))}}])&&y(e.prototype,n),r&&y(e,r),i}(i),S=(w=[],O=function(t){var e=function(t){if(!t)throw new Error("An indentifying class for the element is required");return document.querySelectorAll(".".concat(t,":not(.dicharts-handler--active)"))}(t.className);Array.prototype.forEach.call(e,(function(e){e.classList.add("dicharts-handler--active"),t.plotly&&function(t,e){var n=new E(t,e);n.showLoading(),n.setLayout(e.layout).setConfig(e.config),e.data&&e.data.length?n.setData(e.data).updatePlot():e.csv&&n.csv(e.csv.url).then((function(t){var r;n.setSourceData(t),null===(r=e.csv)||void 0===r||r.onFetch(t,e,n)}))}(e,t.plotly)}))},{addChart:function(t){w.push(t),t.onAdd&&t.onAdd(t),O(t)},getCharts:function(){return w}})}]);
+var DICharts=function(t){var e={};function n(r){if(e[r])return e[r].exports;var o=e[r]={i:r,l:!1,exports:{}};return t[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}return n.m=t,n.c=e,n.d=function(t,e,r){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:r})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var o in t)n.d(r,o,function(e){return t[e]}.bind(null,o));return r},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=1)}([function(t,e,n){"use strict";var r=function(t){return function(t){return!!t&&"object"==typeof t}(t)&&!function(t){var e=Object.prototype.toString.call(t);return"[object RegExp]"===e||"[object Date]"===e||function(t){return t.$$typeof===o}(t)}(t)};var o="function"==typeof Symbol&&Symbol.for?Symbol.for("react.element"):60103;function i(t,e){return!1!==e.clone&&e.isMergeableObject(t)?f((n=t,Array.isArray(n)?[]:{}),t,e):t;var n}function a(t,e,n){return t.concat(e).map((function(t){return i(t,n)}))}function c(t){return Object.keys(t).concat(function(t){return Object.getOwnPropertySymbols?Object.getOwnPropertySymbols(t).filter((function(e){return t.propertyIsEnumerable(e)})):[]}(t))}function u(t,e){try{return e in t}catch(t){return!1}}function l(t,e,n){var r={};return n.isMergeableObject(t)&&c(t).forEach((function(e){r[e]=i(t[e],n)})),c(e).forEach((function(o){(function(t,e){return u(t,e)&&!(Object.hasOwnProperty.call(t,e)&&Object.propertyIsEnumerable.call(t,e))})(t,o)||(u(t,o)&&n.isMergeableObject(e[o])?r[o]=function(t,e){if(!e.customMerge)return f;var n=e.customMerge(t);return"function"==typeof n?n:f}(o,n)(t[o],e[o],n):r[o]=i(e[o],n))})),r}function f(t,e,n){(n=n||{}).arrayMerge=n.arrayMerge||a,n.isMergeableObject=n.isMergeableObject||r,n.cloneUnlessOtherwiseSpecified=i;var o=Array.isArray(e);return o===Array.isArray(t)?o?n.arrayMerge(t,e,n):l(t,e,n):i(e,n)}f.all=function(t,e){if(!Array.isArray(t))throw new Error("first argument should be an array");return t.reduce((function(t,n){return f(t,n,e)}),{})};var s=f;t.exports=s},function(t,e,n){"use strict";n.r(e),n.d(e,"handler",(function(){return S})),n.d(e,"Chart",(function(){return i}));function r(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}function o(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var i=function(){function t(e){var n=this;!function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,t),o(this,"chartElement",void 0),o(this,"theme","default"),o(this,"filters",[]),o(this,"setFilters",(function(t){n.filters=t})),o(this,"getFilters",(function(){return n.filters})),o(this,"showLoading",(function(){var t=n.getParentElement();t&&t.classList.add("chart-container--loading")})),o(this,"hideLoading",(function(){var t=n.getParentElement();t&&function(t){t.classList.remove("chart-container--loading");var e=t.querySelector(".chart-loading");e&&e.classList.add("hide")}(t)})),o(this,"addFilter",(function(t,e,r){var o=n.getElement(t);return e.forEach((function(t){var e=document.createElement("option");e.value=t,e.text=n.getFilterText(t,r),o.appendChild(e)})),o.classList.add("data-selector--active"),o})),o(this,"getThemes",(function(){return{rainbow:["#e84439","#eb642b","#f49b21","#109e68","#0089cc","#893f90","#c2135b","#f8c1b2","#f6bb9d","#fccc8e","#92cba9","#88bae5","#c189bb","#e4819b"],default:["#6c120a","#a21e25","#cd2b2a","#dc372d","#ec6250","#f6b0a0","#fbd7cb","#fce3dc"],sunflower:["#7d4712","#ba6b15","#df8000","#f7a838","#fac47e","#fedcab","#fee7c1","#feedd4"],marigold:["#7a2e05","#ac4622","#cb5730","#ee7644","#f4a57c","#facbad","#fcdbbf","#fde5d4"],rose:["#65093d","#8d0e56","#9f1459","#d12568","#e05c86","#f3a5b6","#f6b8c1","#f9cdd0"],lavendar:["#42184c","#632572","#732c85","#994d98","#af73ae","#cb98c4","#deb5d6","#ebcfe5"],bluebell:["#0a3a64","#00538e","#1060a3","#4397d3","#77adde","#a3c7eb","#bcd4f0","#d3e0f4"],leaf:["#08492f","#005b3e","#00694a","#3b8c62","#74bf93","#a2d1b0","#b1d8bb","#c5e1cb"]}})),o(this,"getTheme",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:"default";return t?[t,n.getThemes()[t]]:[n.theme,n.getThemes()[n.theme]]})),this.chartElement=this.getElement(e)}var e,n,i;return e=t,(n=[{key:"getParentElement",value:function(){return this.chartElement.parentElement}},{key:"getFilterText",value:function(t){var e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},n=e.labelPrefix,r=e.labelSuffix;return"".concat(n||"").concat(t).concat(r||"")}},{key:"handleFilter",value:function(t){var e=this.getParentElement();if(e){var n=e.querySelector(".".concat(t.className));if(!n)throw new Error("No element with class ".concat(t.className));var r=t.options;!r&&t.getOptions&&(r=t.getOptions(this)),r&&r.length&&(this.addFilter(n,r||[]),this.addFilterEvents(n,t),this.setFilters(this.filters.concat({name:t.className,value:[r[0]]})))}}},{key:"addFilterEvents",value:function(t,e){var n=this;e.onChange&&t.addEventListener("change",(function(t){var r=t.target.value;n.setFilters(n.filters.map((function(t){return t.name===e.className&&-1===t.value.indexOf(r)&&(t.value=e.multi?t.value.concat(r):[r]),t}))),e.onChange(t,n)}))}},{key:"initCustomWidgets",value:function(t){var e=this;t.filters&&t.filters.forEach((function(t){return e.handleFilter(t)}))}},{key:"getElement",value:function(t){if(!t)throw new Error("An id to the element or an actual element is required");if("string"==typeof t){var e=document.getElementById(t);if(!e)throw new Error("No element with id ".concat(t));return e}return t}}])&&r(e.prototype,n),i&&r(e,i),t}(),a=n(0),c=n.n(a);function u(t){return(u="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}function l(t,e){return function(t){if(Array.isArray(t))return t}(t)||function(t,e){if("undefined"==typeof Symbol||!(Symbol.iterator in Object(t)))return;var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=t[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!e||n.length!==e);r=!0);}catch(t){o=!0,i=t}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(t,e)||function(t,e){if(!t)return;if("string"==typeof t)return f(t,e);var n=Object.prototype.toString.call(t).slice(8,-1);"Object"===n&&t.constructor&&(n=t.constructor.name);if("Map"===n||"Set"===n)return Array.from(t);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return f(t,e)}(t,e)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function f(t,e){(null==e||e>t.length)&&(e=t.length);for(var n=0,r=new Array(e);n<e;n++)r[n]=t[n];return r}function s(t,e){var n=Object.keys(t);if(Object.getOwnPropertySymbols){var r=Object.getOwnPropertySymbols(t);e&&(r=r.filter((function(e){return Object.getOwnPropertyDescriptor(t,e).enumerable}))),n.push.apply(n,r)}return n}function d(t){for(var e=1;e<arguments.length;e++){var n=null!=arguments[e]?arguments[e]:{};e%2?s(Object(n),!0).forEach((function(e){m(t,e,n[e])})):Object.getOwnPropertyDescriptors?Object.defineProperties(t,Object.getOwnPropertyDescriptors(n)):s(Object(n)).forEach((function(e){Object.defineProperty(t,e,Object.getOwnPropertyDescriptor(n,e))}))}return t}function y(t,e){for(var n=0;n<e.length;n++){var r=e[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(t,r.key,r)}}function h(t,e){return(h=Object.setPrototypeOf||function(t,e){return t.__proto__=e,t})(t,e)}function p(t){var e=function(){if("undefined"==typeof Reflect||!Reflect.construct)return!1;if(Reflect.construct.sham)return!1;if("function"==typeof Proxy)return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],(function(){}))),!0}catch(t){return!1}}();return function(){var n,r=v(t);if(e){var o=v(this).constructor;n=Reflect.construct(r,arguments,o)}else n=r.apply(this,arguments);return b(this,n)}}function b(t,e){return!e||"object"!==u(e)&&"function"!=typeof e?g(t):e}function g(t){if(void 0===t)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return t}function v(t){return(v=Object.setPrototypeOf?Object.getPrototypeOf:function(t){return t.__proto__||Object.getPrototypeOf(t)})(t)}function m(t,e,n){return e in t?Object.defineProperty(t,e,{value:n,enumerable:!0,configurable:!0,writable:!0}):t[e]=n,t}var w,O,j={hovermode:"closest"},P={displayModeBar:!0,responsive:!0,showLink:!1,plotlyServerURL:"https://chart-studio.plotly.com",toImageButtonOptions:{width:1200,height:657},modeBarButtonsToRemove:["pan2d","select2d","lasso2d","hoverClosestCartesian","hoverCompareCartesian","toggleSpikelines"]},E=function(t){!function(t,e){if("function"!=typeof e&&null!==e)throw new TypeError("Super expression must either be null or a function");t.prototype=Object.create(e&&e.prototype,{constructor:{value:t,writable:!0,configurable:!0}}),e&&h(t,e)}(i,t);var e,n,r,o=p(i);function i(t,e){var n;return function(t,e){if(!(t instanceof e))throw new TypeError("Cannot call a class as a function")}(this,i),m(g(n=o.call(this,t)),"options",void 0),m(g(n),"data",[]),m(g(n),"layout",{}),m(g(n),"config",{}),m(g(n),"plot",void 0),m(g(n),"sourceData",void 0),m(g(n),"getPlot",(function(){return n.plot})),m(g(n),"setData",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[];return n.data=t,g(n)})),m(g(n),"setLayout",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return n.layout=c()(n.layout,t),n.setTheme(n.options.theme)})),m(g(n),"getLayout",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return c()(j,t)})),m(g(n),"setTheme",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:"default";return n.layout=d(d({},n.layout),{},{colorway:n.getThemes()[t]}),n.theme=t,g(n)})),m(g(n),"setConfig",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return n.config=c()(n.config,t),g(n)})),m(g(n),"getConfig",(function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return c()(P,t)})),m(g(n),"getOptions",(function(){return n.options})),m(g(n),"csv",(function(t){return new Promise((function(e){window.Plotly.d3.csv(t,(function(t){e(t)}))}))})),m(g(n),"updatePlot",(function(){return n.plot?window.Plotly.react(n.chartElement,n.data,n.getLayout(n.layout),n.getConfig(n.config)).then((function(t){return n.addPlot(t),n.options.theme||n.updateLayoutColorway(),n.hideLoading(),{plot:t,config:n.options}})):n.newPlot(n.chartElement,n.data,n.getLayout(n.layout),n.getConfig(n.config))})),m(g(n),"updateLayoutColorway",(function(){if(n.plot){var t=n.getThemes(),e=void 0;if(n.plot.calcdata.length>8)e=t.rainbow;else for(var r=document.body.classList,o=0,i=Object.entries(t);o<i.length;o++){var a=l(i[o],2),c=a[0],u=a[1];if(r.contains("body--".concat(c))&&u.length<=8){e=u;break}}e&&window.Plotly.relayout(n.chartElement,{colorway:e})}})),n.options=e,n}return e=i,(n=[{key:"setSourceData",value:function(t){return this.sourceData=t,this}},{key:"getSourceData",value:function(){return this.sourceData}},{key:"addPlot",value:function(t){return this.plot=t,this}},{key:"newPlot",value:function(t,e){var n=this,r=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},o=arguments.length>3&&void 0!==arguments[3]?arguments[3]:{};return window.Plotly.newPlot(t,e,r,this.getConfig(o)).then((function(t){return t.on("plotly_click",(function(t){n.options.onClick&&n.options.onClick(t,n)})),!n.plot&&n.options.widgets&&n.initCustomWidgets(n.options.widgets),n.addPlot(t),n.initPlotlyWidgets(),n.options.theme||n.updateLayoutColorway(),n.hideLoading(),{plot:t,config:n.options}}))}},{key:"initPlotlyWidgets",value:function(){var t=this.options.widgets,e=void 0===t?{}:t;e.legend&&this.handleLegend(e.legend)}},{key:"handleLegend",value:function(t){var e=this;this.plot&&t.onClick&&this.plot.on("plotly_legendclick",(function(n){t.onClick&&t.onClick(n,e)}))}}])&&y(e.prototype,n),r&&y(e,r),i}(i),S=(w=[],O=function(t){var e=function(t){if(!t)throw new Error("An indentifying class for the element is required");return document.querySelectorAll(".".concat(t,":not(.dicharts-handler--active)"))}(t.className);!function(t,e){t.plotly&&t.plotly.onAdd&&t.plotly.onAdd(e),t.d3&&t.d3.onAdd&&t.d3.onAdd(e)}(t,e),Array.prototype.forEach.call(e,(function(e){e.classList.add("dicharts-handler--active"),t.plotly&&function(t,e){var n=new E(t,e);n.showLoading(),n.setLayout(e.layout).setConfig(e.config),e.data&&e.data.length?n.setData(e.data).updatePlot():e.csv&&n.csv(e.csv.url).then((function(t){var r;n.setSourceData(t),null===(r=e.csv)||void 0===r||r.onFetch(t,e,n)}))}(e,t.plotly)}))},{addChart:function(t){w.push(t),O(t)},getCharts:function(){return w}})}]);
